+++ examples/data-sources/upcloud_managed_object_storage_policies/data-source.tf
@@ -1,0 +1,9 @@
+resource "upcloud_managed_object_storage" "this" {
+  name              = "example"
+  region            = "europe-1"
+  configured_status = "started"
+}
+
+data "upcloud_Managed_object_storage_policies" "this" {
+  service_uuid = upcloud_managed_object_storage.this.id
+}