+++ internal/service/managedobjectstorage/data_sources.go
@@ -14,7 +14,7 @@ import (
 func DataSourceManagedObjectStorageRegions() *schema.Resource {
 	return &schema.Resource{
 		Description: `Returns a list of available Managed Object Storage regions.`,
-		ReadContext: dataSourceHostsRead,
+		ReadContext: dataSourceRegionsRead,
 		Schema: map[string]*schema.Schema{
 			"regions": {
 				Type:     schema.TypeSet,
@@ -46,7 +46,7 @@ func DataSourceManagedObjectStorageRegions() *schema.Resource {
 	}
 }
 
-func dataSourceHostsRead(ctx context.Context, d *schema.ResourceData, meta interface{}) (diags diag.Diagnostics) {
+func dataSourceRegionsRead(ctx context.Context, d *schema.ResourceData, meta interface{}) (diags diag.Diagnostics) {
 	svc := meta.(*service.Service)
 
 	regions, err := svc.GetManagedObjectStorageRegions(ctx, &request.GetManagedObjectStorageRegionsRequest{})
@@ -79,4 +79,62 @@ func buildManagedObjectStorageRegions(regions []upcloud.ManagedObjectStorageRegi
 	}
 
 	return maps
+}
+
+func DataSourceManagedObjectStoragePolicies() *schema.Resource {
+	return &schema.Resource{
+		Description: "Policies available for a Managed Object Storage resource. See `managed_object_storage_user_policy` for attaching to a user.",
+		ReadContext: dataSourcePoliciesRead,
+		Schema: map[string]*schema.Schema{
+			"policies": {
+				Description: "Policies.",
+				Type:        schema.TypeSet,
+				Computed:    true,
+				Elem: &schema.Resource{
+					Schema: schemaPolicy(),
+				},
+			},
+			"service_uuid": {
+				Description: "Service UUID.",
+				Required:    true,
+				Type:        schema.TypeString,
+			},
+		},
+	}
+}
+
+func dataSourcePoliciesRead(ctx context.Context, d *schema.ResourceData, meta interface{}) (diags diag.Diagnostics) {
+	svc := meta.(*service.Service)
+	serviceUUID := d.Get("service_uuid").(string)
+
+	policies, err := svc.GetManagedObjectStoragePolicies(ctx, &request.GetManagedObjectStoragePoliciesRequest{ServiceUUID: serviceUUID})
+	if err != nil {
+		return diag.FromErr(err)
+	}
+
+	d.SetId(serviceUUID)
+
+	return setManagedObjectStoragePoliciesData(d, policies)
+}
+
+func setManagedObjectStoragePoliciesData(d *schema.ResourceData, policies []upcloud.ManagedObjectStoragePolicy) (diags diag.Diagnostics) {
+	policyMaps := make([]map[string]interface{}, 0)
+	for _, policy := range policies {
+		policyMaps = append(policyMaps, map[string]interface{}{
+			"arn":                policy.Arn,
+			"attachment_count":   policy.AttachmentCount,
+			"created_at":         policy.CreatedAt.String(),
+			"default_version_id": policy.DefaultVersionId,
+			"description":        policy.Description,
+			"document":           policy.Document,
+			"name":               policy.Name,
+			"system":             policy.System,
+			"updated_at":         policy.UpdatedAt,
+		})
+	}
+	if err := d.Set("policies", policyMaps); err != nil {
+		return diag.FromErr(err)
+	}
+
+	return diags
 }