+++ CHANGELOG.md
@@ -5,6 +5,12 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [5.0.1] - 2024-03-01
+
+### Fixed
+
+- Added missing data sources and resources to Terraform provider documentation
+
 ## [5.0.0] - 2024-02-29
 
 ### Added
@@ -505,7 +511,8 @@ Updated upcloud-go-api, added build/CI scripts, and repackaged 0.1.0 as 1.0.0.
 - resource_upcloud_firewall_rule removed and replaced by resource_upcloud_firewall_rules
 - resource_upcloud_zone removed and replaced by zone and zones datasources
 
-[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.0.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.0.1...HEAD
+[5.0.1]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.0.0...v5.0.1
 [5.0.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v4.1.0...v5.0.0
 [4.1.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v4.0.0...v4.1.0
 [4.0.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v3.4.0...v4.0.0